# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Ewura Hub API - Airtime & Data selling platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user or agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /orders:
    get:
      operationId: getOrders
      tags: [orders]
      summary: Get orders (current user's or all for admin)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, processing, completed, failed]
        - name: network
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrdersResponse"

    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderWithPayment"
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get a single order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/orders:
    get:
      operationId: getAdminOrders
      tags: [admin]
      summary: Get all orders (admin)
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: network
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: All orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrdersResponse"
  /admin/agents:
    get:
      operationId: getAdminAgents
      tags: [admin]
      summary: Get all agents for admin
      parameters:
        - name: verified
          in: query
          schema:
            type: boolean
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Agents list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentsResponse"
  /admin/agents/{id}/verify:
    patch:
      operationId: verifyAgent
      tags: [admin]
      summary: Verify or unverify an agent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyAgentBody"
      responses:
        "200":
          description: Agent updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /admin/agents/{id}/balance:
    patch:
      operationId: adjustAgentBalance
      tags: [admin]
      summary: Adjust agent wallet balance
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdjustBalanceBody"
      responses:
        "200":
          description: Balance adjusted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wallet"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
...
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [user, agent, admin]
        isVerified:
          type: boolean
        walletBalance:
          type: number
        createdAt:
          type: string
          format: date-time
      required: [id, username, email, phone, role, isVerified, walletBalance, createdAt]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        message:
          type: string
      required: [user, message]
    RegisterBody:
      type: object
      properties:
        username:
          type: string
          minLength: 3
        email:
          type: string
          format: email
        phone:
          type: string
        password:
          type: string
          minLength: 6
        role:
          type: string
          enum: [user, agent]
          default: user
      required: [username, email, phone, password]
    LoginBody:
      type: object
      properties:
        emailOrPhone:
          type: string
        password:
          type: string
      required: [emailOrPhone, password]
...
    Order:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        username:
          type: string
        network:
          type: string
          enum: [MTN, Telecel, AirtelTigo]
        type:
          type: string
          enum: [airtime, data]
        productName:
          type: string
        recipientPhone:
          type: string
        amount:
          type: number
        status:
          type: string
          enum: [pending, processing, completed, failed]
        paymentMethod:
          type: string
          enum: [paystack, wallet]
        paymentReference:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, userId, username, network, type, productName, recipientPhone, amount, status, paymentMethod, createdAt]
    OrderWithPayment:
      type: object
      properties:
        order:
          $ref: "#/components/schemas/Order"
        paymentUrl:
          type: string
        reference:
          type: string
        message:
          type: string
      required: [order, message]
    OrdersResponse:
      type: object
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        page:
          type: integer
        pages:
          type: integer
      required: [orders, total, page, pages]
    CreateOrderBody:
      type: object
      properties:
        productId:
          type: string
        recipientPhone:
          type: string
        paymentMethod:
          type: string
          enum: [paystack, wallet]
      required: [productId, recipientPhone, paymentMethod]
    Wallet:
      type: object
      properties:
        balance:
          type: number
        userId:
          type: string
        totalFunded:
          type: number
        totalSpent:
          type: number
      required: [balance, userId, totalFunded, totalSpent]
...
    AgentsResponse:
      type: object
      properties:
        agents:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        pages:
          type: integer
      required: [agents, total, page, pages]
    VerifyAgentBody:
      type: object
      properties:
        isVerified:
          type: boolean
      required: [isVerified]
    AdjustBalanceBody:
      type: object
      properties:
        amount:
          type: number
        type:
          type: string
          enum: [credit, debit]
        reason:
          type: string
      required: [amount, type, reason]